+++ app/routes.ts
@@ -1,5 +1,5 @@
-import { index } from '@react-router/dev/routes'
+import { index } from "@react-router/dev/routes";
 
-import type { RouteConfig } from '@react-router/dev/routes'
+import type { RouteConfig } from "@react-router/dev/routes";
 
-export default [index('routes/home.tsx')] satisfies RouteConfig
+export default [index("routes/home.tsx")] satisfies RouteConfig;